backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels_redis==4.2.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==4.2.16
django-cors-headers==4.6.0
django-environ==0.11.2
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwcrypto==1.5.6
msgpack==1.1.0
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
sqlparse==0.5.1
tomli==2.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.1.1
psycopg2==2.9.3